Adds Drag&Drop working and tested!
First I created an image, then selected a sprite to use (an image called Icon and another one called Slot on the hierarchy - we can just replace the sprite to use the ones we want). The order of these in the hierarchy matters: first Canvas -> Panel -> Slot -> Icon. Work and tested on RuneScene.